enforced custom permissions to mutations

+++ api/everecon/schema_event.py
@@ -2,12 +2,13 @@ from random import sample
 
 import graphene
 from django.contrib.auth.models import User
-from django_graphene_permissions import permissions_checker
+from django_graphene_permissions import permissions_checker, check_object_permissions, PermissionDenied
 from django_graphene_permissions.permissions import IsAuthenticated
 from graphene_django import DjangoObjectType
 
 from .email_sendgrid import *
 from .models import Category, Community, Event, Speaker, Tag
+from .permissions import *
 from .schema_community import CategoryType, CommunityType, EventType, TagType
 
 
@@ -32,10 +33,12 @@ class CreateEvent(graphene.Mutation):
     tags = graphene.List(TagType)
     category = graphene.Field(CategoryType)
 
-    @permissions_checker([IsAuthenticated])
+    @permissions_checker([IsCoreMember])
     def mutate(root, info, **kwargs):
         # print(kwargs)
         community = Community.objects.get(id=kwargs.pop("community"))
+        if not check_object_permissions([IsCoreMember], info.context, community):
+            raise PermissionDenied()
         category = Category.objects.get(id=kwargs.pop("category"))
         # tags = Tag.objects.filter(id__in=kwargs.pop("tags"))
         # tag_obj =
@@ -95,12 +98,15 @@ class Checkin4Event(graphene.Mutation):
     ok = graphene.Boolean()
     message = graphene.String()
 
-    @permissions_checker([IsAuthenticated])
+    @permissions_checker([IsVolunteer])
     def mutate(root, info, **kwargs):
         eventid = kwargs.pop("eventid")
         userid = kwargs.pop("userid")
         user = User.objects.get(id=userid)
-        attends = Event.objects.get(id=eventid).attendees.all()
+        event = Event.objects.get(id=eventid)
+        if not check_object_permissions([IsVolunteer], info.context, event):
+            raise PermissionDenied()
+        attends = event.attendees.all()
         if user in attends:
             Event.objects.get(id=eventid).checkins.add(user)
             ok = True
@@ -119,12 +125,14 @@ class Uncheck4Event(graphene.Mutation):
     ok = graphene.Boolean()
     message = graphene.String()
 
-    @ permissions_checker([IsAuthenticated])
+    @ permissions_checker([IsVolunteer])
     def mutate(root, info, *args, **kwargs):
         eventid = kwargs.pop("eventid")
         userid = kwargs.pop("userid")
         user = User.objects.get(id=userid)
         event = Event.objects.get(id=eventid)
+        if not check_object_permissions([IsVolunteer], info.context, event):
+            raise PermissionDenied()
         # user = User.objects.get(kwargs.get(''))
         event.checkins.remove(user)
         return Uncheck4Event(ok=True, message="Checkin removed")
@@ -137,9 +145,11 @@ class AddSpeaker(graphene.Mutation):
 
     ok = graphene.Boolean() 
 
-    @ permissions_checker([IsAuthenticated])
+    @ permissions_checker([IsCoreMember])
     def mutate(root, info, eventid, speakerid):
         event = Event.objects.get(id=eventid)
+        if not check_object_permissions([IsCoreMember], info.context, event):
+            raise PermissionDenied()
         # user = User.objects.get(kwargs.get(''))
         event.speakers.add(speakerid)
         speaker = Speaker.objects.get(id=speakerid)
@@ -154,9 +164,11 @@ class RemoveSpeaker(graphene.Mutation):
 
     ok = graphene.Boolean()
 
-    @ permissions_checker([IsAuthenticated])
+    @ permissions_checker([IsCoreMember])
     def mutate(root, info, eventid, speakerid):
         event = Event.objects.get(id=eventid)
+        if not check_object_permissions([IsCoreMember], info.context, event):
+            raise PermissionDenied()
         # user = User.objects.get(kwargs.get(''))
         event.speakers.remove(speakerid)
         return RemoveSpeaker(ok=True)
@@ -184,7 +196,7 @@ class UpdateEvent(graphene.Mutation):
     tags = graphene.List(TagType)
     category = graphene.Field(CategoryType)
 
-    @ permissions_checker([IsAuthenticated])
+    @ permissions_checker([IsCoreMember])
     def mutate(root, info, **kwargs):
         id = kwargs.pop("id")
         try:
@@ -192,6 +204,11 @@ class UpdateEvent(graphene.Mutation):
 
         except Exception:
             pass
+
+        event = Event.objects.get(id=id)
+        if not check_object_permissions([IsCoreMember], info.context, event):
+            raise PermissionDenied()
+
         Event.objects.filter(id=id).update(**kwargs)
         event = Event.objects.get(id=id)
         if tags:
@@ -232,9 +249,11 @@ class DeleteEvent(graphene.Mutation):
 
     ok = graphene.Boolean()
 
-    @ permissions_checker([IsAuthenticated])
+    @ permissions_checker([IsCoreMember])
     def mutate(root, info, **kwargs):
         obj = Event.objects.get(pk=kwargs["id"])
+        if not check_object_permissions([IsCoreMember], info.context, obj):
+            raise PermissionDenied()
         obj.delete()
         return DeleteEvent(ok=True)
 
@@ -248,11 +267,14 @@ class UpdateEventImage(graphene.Mutation):
     success = graphene.String()
     picture = graphene.String()
 
+    @permissions_checker([IsCoreMember])
     def mutate(self,  info, id, *args, **kwargs):
         # When using it in Django, context will be the request
         files = info.context.FILES
         event: Event
         event = Event.objects.get(id=id)
+        if not check_object_permissions([IsCoreMember], info.context, event):
+            raise PermissionDenied()
         event.featured_image = files["file"]
         event.save()
         event = Event.objects.get(id=id)
@@ -298,7 +320,6 @@ class Query(graphene.ObjectType):
         return Category.objects.get(id=cat).event_set.all()
 
     def resolve_categories(root, info):
-        print("hello")
         return Category.objects.all()
